requirements.txt
scipy
requests
python-firebase